remove old stock page
