--- src/core/Adapters/Stocks/CompanyProfile.cs
@@ -1,8 +1,0 @@
-namespace core.Adapters.Stocks
-{
-	public class CompanyProfile
-	{
-		public string Ticker { get; set; }
-		public CompanyProfileDetails Profile { get; set; }
-	}
-}
