+++ src/core/Adapters/Stocks/IStocksService2.cs
@@ -4,7 +4,7 @@ namespace core.Adapters.Stocks
 {
     public interface IStocksService2
     {
-        Task<CompanyProfile2> GetCompanyProfile(string ticker);
+        Task<CompanyProfile> GetCompanyProfile(string ticker);
         Task<StockAdvancedStats> GetAdvancedStats(string ticker);
         Task<TickerPrice> GetPrice(string ticker);
     }
