+++ src/web/ClientApp/src/app/stock-details/stock-details.component.css
@@ -2,3 +2,11 @@ label {
 	font-weight: bold;
 	margin-top: 10px;
 }
+
+.stat-label {
+  font-weight: 500;
+}
+
+div.row {
+  margin-top: 10px;
+}
