+++ src/web/ClientApp/src/app/stock-details/stock-details.component.html
@@ -1,25 +1,70 @@
 <div *ngIf="loaded" class="top-container">
 
+  <a href="https://robinhood.com/stocks/{{stock.ticker}}" class="btn dashboard-button" target="_blank">
+    <img src="robinhood.png" height="30px" />
+  </a>
 	<button class="btn btn-secondary dashboard-button" (click)="buyStock()">Buy</button>
 
 	<div>
-		<label>{{stock.profile.companyName}}</label> - {{stock.profile.industry}}/{{stock.profile.sector}}
-		<div>
-			Ticker {{stock.ticker}}, <a href="https://robinhood.com/stocks/{{stock.ticker}}" target="_blank">link on robinhood</a> <br />
+    <img style="float: left" src="https://storage.googleapis.com/iex/api/logos/{{profile.symbol}}.png" />
+    <h3>{{profile.symbol}}: {{profile.companyName}}</h3>
+    <i>{{profile.industry}} / {{profile.sector}}</i>
+    <div>
+      <a href="{{profile.website}}" target="_blank">{{profile.website}}</a>
     </div>
-	</div>
-	<div>
-		<label>Price information</label>
-		<div>
-			Priced at <b>${{stock.price}}</b> with largest gain of {{stock.largestGain}}% and largest loss of {{stock.largestLoss}}%
-		</div>
-	</div>
+    <div>
+      <i>{{profile.description}}</i>
+    </div>
+  </div>
 
-	<div>
-		<label>Ratios</label>
-		<div>
-			Book value of {{stock.bookValue}} with P/E ratio of {{stock.peValue}} (information is {{stock.age}} months old)
-		</div>
+	<div style="clear: both">
+		<div class="container">
+      <div class="row">
+        <div class="col-12" style="text-align: center;"><h4>Fundamentals</h4></div>
+      </div>
+      <div class="row">
+        <div class="col"><div class="stat-label">Price</div><div>{{stock.price | currency}}</div></div>
+        <div class="col"><div class="stat-label">PE</div><div>{{stats.peRatio}}</div></div>
+        <div class="col"><div class="stat-label">10d Volume</div><div>{{stats.avg10Volume / 1000000 | number}}M</div></div>
+        <div class="col"><div class="stat-label">30d Volumne</div><div>{{stats.avg30Volume / 1000000 | number}}M</div></div>
+      </div>
+      <div class="row">
+        <div class="col"><div class="stat-label">1 Year High</div><div>{{stats.week52highDate | date}}</div></div>
+        <div class="col"><div class="stat-label">1 Year High</div><div>{{stats.week52High | currency}}</div></div>
+        <div class="col"><div class="stat-label">1 Year Low</div><div>{{stats.week52lowDate | date}}</div></div>
+        <div class="col"><div class="stat-label">1 Year Low</div><div>{{stats.week52Low | currency}}</div></div>
+      </div>
+      <div class="row">
+        <div class="col"><div class="stat-label">Next Earnings</div><div>{{stats.nextEarningsDate | date}}</div></div>
+        <div class="col"><div class="stat-label">Ex Div Date</div><div>{{stats.exDividendDate | date}}</div></div>
+        <div class="col"><div class="stat-label">Next Div Date</div><div>{{stats.nextDividendDate | date}}</div></div>
+        <div class="col"><div class="stat-label">Dividend Yield</div><div>{{stats.dividendYield | number}}</div></div>
+      </div>
+      <div class="row">
+        <div class="col"><div class="stat-label">Market Cap</div><div>{{stats.marketCap / 1000000 | number}}M</div></div>
+        <div class="col"><div class="stat-label">Beta</div><div>{{stats.beta | number}}</div></div>
+        <div class="col"><div class="stat-label">Put/Call Ratio</div><div>{{stats.putCallRatio | number}}</div></div>
+        <div class="col"><div class="stat-label">Price/Book</div><div>{{stats.priceToBook | number}}</div></div>
+      </div>
+      <div class="row">
+        <div class="col"><div class="stat-label">Revenue</div><div>{{stats.revenue / 1000000 | number}}M</div></div>
+        <div class="col"><div class="stat-label">Gross Profit</div><div>{{stats.grossProfit / 1000000 | number}}M</div></div>
+        <div class="col"><div class="stat-label">Profit Margin</div><div>{{stats.profitMargin | percent}}</div></div>
+        <div class="col"><div class="stat-label">Total Cash</div><div>{{stats.totalCash / 1000000 | number}}M</div></div>
+      </div>
+      <div class="row">
+        <div class="col"><div class="stat-label">5-year change</div><div>{{stats.year5ChangePercent | percent}}</div></div>
+        <div class="col"><div class="stat-label">1-year change</div><div>{{stats.year1ChangePercent | percent}}</div></div>
+        <div class="col"><div class="stat-label">3-month change</div><div>{{stats.month3ChangePercent | percent}}</div></div>
+        <div class="col"><div class="stat-label">1-month change</div><div>{{stats.month1ChangePercent | percent}}</div></div>
+      </div>
+      <div class="row">
+        <div class="col"><div class="stat-label">Country</div><div>{{profile.country}}</div></div>
+        <div class="col"><div class="stat-label">Type</div><div>{{profile.issueType}}</div></div>
+        <div class="col"><div class="stat-label">Employees</div><div>{{profile.employees}}</div></div>
+        <div class="col"></div>
+      </div>
+
 	</div>
 
 	<div>
@@ -28,7 +73,7 @@
 			<google-chart style="width: 100%;" [type]="priceChartType" [data]="priceChartData" [options]="priceOptions"></google-chart>
 		</div>
 		<div>
-			<google-chart style="width: 100%;" [type]="priceChartType" [data]="volumeChartData" [options]="volumeOptions"></google-chart>
+			<google-chart style="width: 100%;" [type]="volumeChartType" [data]="volumeChartData" [options]="volumeOptions"></google-chart>
 		</div>
 		<div>
 			<google-chart style="width: 100%;" [type]="priceChartType" [data]="bookChartData" [options]="bookOptions"></google-chart>
