--- src/web/ClientApp/src/app/stock-details2/stock-details.component.css
@@ -1,12 +1,0 @@
-label {
-	font-weight: bold;
-	margin-top: 10px;
-}
-
-.stat-label {
-  font-weight: 500;
-}
-
-div.row {
-  margin-top: 10px;
-}
